refactor(core): extract PhaseRunner from Executor
Move phase execution, parallel/sequential dispatch, and recovery logic
into a dedicated PhaseRunner class with dependency injection, reducing
executor.ts from 1437 to 1079 lines.

Closes #113
